% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{clean_string}
\alias{clean_string}
\title{Clean a character string}
\usage{
clean_string(x)
}
\arguments{
\item{x}{A character string to be cleaned.}
}
\value{
A cleaned character string.
}
\description{
This function trims leading and trailing whitespace and converts
the string to lowercase.
}
\examples{
clean_string("  Hello World!  ")
}
