% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fetch_sample_data}
\alias{fetch_sample_data}
\title{Fetch sample data}
\usage{
fetch_sample_data()
}
\value{
A data frame with sample data.
}
\description{
This function returns a small example data frame with sample data.
}
\examples{
fetch_sample_data()
}
